Insights data storage completed // * todo         // get and set total activities for the week         // get and set the percentage differences and set the icon         // handle when there is no data         // plot line chart for the week

+++ app/src/main/java/tech/ayodele/gravity/UserDetails.kt
@@ -1,9 +1,7 @@
 package tech.ayodele.gravity
 
-import android.os.Parcel
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
-import java.io.Serializable
 
 @Parcelize // for sending the user details as object across activities
 data class UserDetails(
@@ -12,7 +10,7 @@ data class UserDetails(
     val email: String? = null,
     val password: String? = null,
     val height: Int? = null,
-    val weight: Int? = null,
+    val weight: Double? = null,
     val gpName: String? = null,
     val gpHospital: String? = null
 ): Parcelable
